dom/future/tests/test_bug883683.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Future - bug 883683</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript"><!--

function runTest() {
  [{}, {}, {}, {}, {}].reduce(Future.reject);
  ok(true, "No leaks with reject?");

  [{}, {}, {}, {}, {}].reduce(Future.resolve);
  ok(true, "No leaks with resolve?");

  [{}, {}, {}, {}, {}].reduce(function(a, b, c, d) { return new Future(function(r) { throw a; }); });
  ok(true, "No leaks with exception?");

  [{}, {}, {}, {}, {}].reduce(function(a, b, c, d) { return new Future(function(r) { }); });
  ok(true, "No leaks with empty future?");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [["dom.future.enabled", true]]}, runTest);
// -->
</script>
</pre>
</body>
</html>
